+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallPlacement.swift
@@ -562,9 +562,19 @@ enum InternalSuperwallPlacement {
         return .transactionTimeout(paywallInfo: paywallInfo)
       }
     }
+    enum StoreKitVersion: String {
+      case storeKit1 = "STOREKIT1"
+      case storeKit2 = "STOREKIT2"
+    }
+    enum TransactionSource: String {
+      case `internal` = "SUPERWALL"
+      case external = "APP"
+    }
     let paywallInfo: PaywallInfo
     let product: StoreProduct?
     let model: StoreTransaction?
+    let source: TransactionSource
+    let storeKitVersion: StoreKitVersion
 
     var audienceFilterParams: [String: Any] {
       switch state {
@@ -579,7 +589,9 @@ enum InternalSuperwallPlacement {
 
     func getSuperwallParameters() async -> [String: Any] {
       var placementParams: [String: Any] = [
-        "store": "APP_STORE"
+        "store": "APP_STORE",
+        "source": source.rawValue,
+        "storeKitVersion": storeKitVersion.rawValue
       ]
 
       switch state {