+++ Sources/SuperwallKit/Paywall/Presentation/PaywallInfo.swift
@@ -125,6 +125,9 @@ public final class PaywallInfo: NSObject {
   /// Information about the presentation of the paywall.
   public let presentation: PaywallPresentationInfo
 
+  /// Indicates whether scrolling of the webview is enabled.
+  public let isScrollEnabled: Bool
+
   init(
     databaseId: String,
     identifier: String,
@@ -155,7 +158,8 @@ public final class PaywallInfo: NSObject {
     localNotifications: [LocalNotification],
     computedPropertyRequests: [ComputedPropertyRequest],
     surveys: [Survey],
-    presentation: PaywallPresentationInfo
+    presentation: PaywallPresentationInfo,
+    isScrollEnabled: Bool
   ) {
     self.databaseId = databaseId
     self.identifier = identifier
@@ -221,6 +225,7 @@ public final class PaywallInfo: NSObject {
     self.shimmerLoadCompleteTime = shimmerLoadCompleteTime?.isoString ?? ""
 
     self.closeReason = closeReason
+    self.isScrollEnabled = isScrollEnabled
   }
 
   func placementParams(
@@ -252,7 +257,8 @@ public final class PaywallInfo: NSObject {
       "variant_id": experiment?.variant.id as Any,
       "cache_key": cacheKey,
       "build_id": buildId,
-      "close_reason": closeReason.description
+      "close_reason": closeReason.description,
+      "is_scroll_enabled": isScrollEnabled as Any
     ]
 
     var loadingVars: [String: Any] = [:]
@@ -359,7 +365,8 @@ extension PaywallInfo: Stubbable {
       presentation: .init(
         style: .none,
         delay: 0
-      )
+      ),
+      isScrollEnabled: true
     )
   }
 
@@ -406,7 +413,8 @@ extension PaywallInfo: Stubbable {
       presentation: .init(
         style: .none,
         delay: 0
-      )
+      ),
+      isScrollEnabled: true
     )
   }
 }