+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -319,6 +319,8 @@ public class PaywallViewController: UIViewController, LoadingDelegate {
         await webView.loadURL(from: paywall)
       }
     }
+
+    webView.scrollView.isScrollEnabled = paywall.isScrollEnabled
   }
 
   private func loadWebViewFromArchive(url: URL) {