+++ Sources/SuperwallKit/StoreKit/Transactions/TransactionManager.swift
@@ -68,6 +68,14 @@ final class TransactionManager {
 
     let result = await purchase(product, purchaseSource: purchaseSource)
 
+    // Return early if using a purchase controller and purchasing externally.
+    // This avoids duplicate calls by the purchase function of the purchase
+    // controller.
+    if case .external = purchaseSource,
+      factory.makeHasExternalPurchaseController() {
+      return result
+    }
+
     switch result {
     case .purchased:
       await didPurchase(
@@ -276,7 +284,9 @@ final class TransactionManager {
         state: .restore(restoreType),
         paywallInfo: paywallInfo,
         product: product,
-        model: transaction
+        model: transaction,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(transactionRestore)
       await paywallViewController.webView.messageHandler.handle(.transactionRestore)
@@ -290,7 +300,9 @@ final class TransactionManager {
         state: .restore(restoreType),
         paywallInfo: .empty(),
         product: product,
-        model: transaction
+        model: transaction,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -301,7 +313,8 @@ final class TransactionManager {
     purchaseSource: PurchaseSource
   ) async -> PurchaseResult {
     await factory.makePurchasingCoordinator().beginPurchase(
-      of: product.productIdentifier
+      of: product.productIdentifier,
+      source: purchaseSource
     )
     switch purchaseSource {
     case .internal:
@@ -341,7 +354,9 @@ final class TransactionManager {
           state: .fail(.failure(error.safeLocalizedDescription, product)),
           paywallInfo: paywallInfo,
           product: product,
-          model: nil
+          model: nil,
+          source: .internal,
+          storeKitVersion: .storeKit1
         )
         await Superwall.shared.track(trackedEvent)
         await paywallViewController.webView.messageHandler.handle(.transactionFail)
@@ -362,7 +377,9 @@ final class TransactionManager {
           state: .fail(.failure(error.safeLocalizedDescription, product)),
           paywallInfo: .empty(),
           product: product,
-          model: nil
+          model: nil,
+          source: .external,
+          storeKitVersion: .storeKit1
         )
         await Superwall.shared.track(trackedEvent)
       }
@@ -390,7 +407,9 @@ final class TransactionManager {
         state: .start(product),
         paywallInfo: paywallInfo,
         product: product,
-        model: nil
+        model: nil,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(transactionStart)
       await paywallViewController.webView.messageHandler.handle(.transactionStart)
@@ -399,6 +418,11 @@ final class TransactionManager {
         paywallViewController.loadingState = .loadingPurchase
       }
     case .external:
+      // If an external purchase controller is being used, skip because this will
+      // get called by the purchase function of the purchase controller.
+      if factory.makeHasExternalPurchaseController() {
+        return
+      }
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -409,7 +433,9 @@ final class TransactionManager {
         state: .start(product),
         paywallInfo: .empty(),
         product: product,
-        model: nil
+        model: nil,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -504,7 +530,9 @@ final class TransactionManager {
         state: .abandon(product),
         paywallInfo: paywallInfo,
         product: product,
-        model: nil
+        model: nil,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(transactionAbandon)
       await paywallViewController.webView.messageHandler.handle(.transactionAbandon)
@@ -525,7 +553,9 @@ final class TransactionManager {
         state: .abandon(product),
         paywallInfo: .empty(),
         product: product,
-        model: nil
+        model: nil,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -548,7 +578,9 @@ final class TransactionManager {
         state: .fail(.pending("Needs parental approval")),
         paywallInfo: paywallInfo,
         product: nil,
-        model: nil
+        model: nil,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(transactionFail)
       await paywallViewController.webView.messageHandler.handle(.transactionFail)
@@ -564,7 +596,9 @@ final class TransactionManager {
         state: .fail(.pending("Needs parental approval")),
         paywallInfo: .empty(),
         product: nil,
-        model: nil
+        model: nil,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
@@ -627,7 +661,9 @@ final class TransactionManager {
         state: .complete(product, transaction),
         paywallInfo: paywallInfo,
         product: product,
-        model: transaction
+        model: transaction,
+        source: .internal,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(transactionComplete)
       await paywallViewController.webView.messageHandler.handle(.transactionComplete)
@@ -667,7 +703,9 @@ final class TransactionManager {
         state: .complete(product, transaction),
         paywallInfo: .empty(),
         product: product,
-        model: transaction
+        model: transaction,
+        source: .external,
+        storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(transactionComplete)